python/requirements.txt
bleach==3.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cryptography==3.3.1
idna==2.10
importlib-metadata==3.1.1
keyring==21.5.0
packaging==20.8
pkginfo==1.6.1
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.3
PyJWT==1.7.1
pyparsing==2.4.7
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
tqdm==4.54.1
urllib3==1.26.2
webencodings==0.5.1
zipp==3.4.0
authlib==0.15.2